# src/main/java/com/adarsh/zorvyn/Request/RegisterUserRequest.java
package com.adarsh.zorvyn.Request;

import com.adarsh.zorvyn.Entity.Role;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class RegisterUserRequest {
    @NotBlank(message = "Username is required")
    private String username;

    @Email(message = "Invalid email format")
    private String email;

    @NotBlank(message = "Password is required")
    private String password;

    @NotNull(message = "Role is required it can be of type (ADMIN, ANALYST, or VIEWER)")
    private Role role;
}